bug fixes Can go to next options in first file without needing to complete notes section (it goes not responsive when I click next when nothing is put)
Page 2 volume of sample and volume of 2xrsob to be in whole numbers without decimal points
e.g. instead of 877.77 it should be 878 or instead of 123.23 it should be 123
enter issue fix is possible but not a need it can be done later

page 3 lane contents
That to be filled in an excel manner where I can put all the data and copy it etc. not row by row at all,
but a table where I can paste all the data plus in the excel sheet it had only 3 columns and it to have 3
columns and to us to enter any data we want, not numbers or specific conditions to follow to be able to
enter as it changes , for instance in sample name and concentration we put eg. BCM (2ug) to enter
things like that